client/requirements.txt
pygame==2.5.2
numpy>=1.24
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
